Consolidate .netcoreapp.cs files because System.Threading.* projects is no longer cross-compiled
