+++ src/libraries/System.Threading.Tasks/tests/Task/TaskStatusTest.cs
@@ -3,11 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using Xunit;
-using System;
 using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
 using System.Diagnostics;
+using System.Runtime.CompilerServices;
 
 namespace System.Threading.Tasks.Tests.Status
 {
@@ -471,6 +469,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         public static void TaskStatus5()
         {
@@ -482,6 +481,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         public static void TaskStatus6()
         {
@@ -492,6 +492,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         public static void TaskStatus7()
         {
@@ -529,6 +530,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         public static void TaskStatus10()
         {
@@ -553,6 +555,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         [OuterLoop]
         public static void TaskStatus12()
@@ -567,6 +570,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         [OuterLoop]
         public static void TaskStatus13()
@@ -580,6 +584,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         public static void TaskStatus14()
         {
@@ -593,6 +598,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         public static void TaskStatus15()
         {
@@ -605,6 +611,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         [OuterLoop]
         public static void TaskStatus16()
@@ -619,6 +626,7 @@ namespace System.Threading.Tasks.Tests.Status
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
         }
+
         [Fact]
         public static void TaskStatus17()
         {
@@ -631,6 +639,99 @@ namespace System.Threading.Tasks.Tests.Status
             };
             TaskStatusTest test = new TaskStatusTest(parameters);
             test.RealRun();
+        }
+
+        [Theory]
+        [MemberData(nameof(Status_IsProperties_Match_MemberData))]
+        public void Status_IsProperties_Match(StrongBox<Task> taskBox)
+        {
+            // The StrongBox<Task> is a workaround for xunit trying to automatically
+            // Dispose of any IDisposable passed into a theory, but Task doesn't like
+            // being Dispose'd when it's not in a final state.
+            Task task = taskBox.Value;
+
+            if (task.IsCompletedSuccessfully)
+            {
+                Assert.Equal(TaskStatus.RanToCompletion, task.Status);
+            }
+            else if (task.IsFaulted)
+            {
+                Assert.Equal(TaskStatus.Faulted, task.Status);
+            }
+            else if (task.IsCanceled)
+            {
+                Assert.Equal(TaskStatus.Canceled, task.Status);
+            }
+
+            switch (task.Status)
+            {
+                case TaskStatus.RanToCompletion:
+                    Assert.True(task.IsCompleted, "Expected IsCompleted to be true");
+                    Assert.True(task.IsCompletedSuccessfully, "Expected IsCompletedSuccessfully to be true");
+                    Assert.False(task.IsFaulted, "Expected IsFaulted to be false");
+                    Assert.False(task.IsCanceled, "Expected IsCanceled to be false");
+                    break;
+
+                case TaskStatus.Faulted:
+                    Assert.True(task.IsCompleted, "Expected IsCompleted to be true");
+                    Assert.False(task.IsCompletedSuccessfully, "Expected IsCompletedSuccessfully to be false");
+                    Assert.True(task.IsFaulted, "Expected IsFaulted to be true");
+                    Assert.False(task.IsCanceled, "Expected IsCanceled to be false");
+                    break;
+
+                case TaskStatus.Canceled:
+                    Assert.True(task.IsCompleted, "Expected IsCompleted to be true");
+                    Assert.False(task.IsCompletedSuccessfully, "Expected IsCompletedSuccessfully to be false");
+                    Assert.False(task.IsFaulted, "Expected IsFaulted to be false");
+                    Assert.True(task.IsCanceled, "Expected IsCanceled to be true");
+                    break;
+
+                default:
+                    Assert.False(task.IsCompleted, "Expected IsCompleted to be false");
+                    Assert.False(task.IsCompletedSuccessfully, "Expected IsCompletedSuccessfully to be false");
+                    Assert.False(task.IsFaulted, "Expected IsFaulted to be false");
+                    Assert.False(task.IsCanceled, "Expected IsCanceled to be false");
+                    break;
+            }
+        }
+
+        public static IEnumerable<object[]> Status_IsProperties_Match_MemberData()
+        {
+            yield return new object[] { new StrongBox<Task>(Task.CompletedTask) };
+
+            yield return new object[] { new StrongBox<Task>(new Task(() => { })) };
+
+            yield return new object[] { new StrongBox<Task>(new TaskCompletionSource<int>().Task) };
+
+            {
+                var tcs = new TaskCompletionSource<int>();
+                tcs.SetResult(42);
+                yield return new object[] { new StrongBox<Task>(tcs.Task) };
+            }
+
+            {
+                var tcs = new TaskCompletionSource<int>();
+                tcs.SetException(new Exception());
+                yield return new object[] { new StrongBox<Task>(tcs.Task) };
+            }
+
+            {
+                var tcs = new TaskCompletionSource<int>();
+                tcs.SetCanceled();
+                yield return new object[] { new StrongBox<Task>(tcs.Task) };
+            }
+
+            {
+                var t = Task.Run(() => { });
+                t.Wait();
+                yield return new object[] { new StrongBox<Task>(t) };
+            }
+
+            {
+                var atmb = new AsyncTaskMethodBuilder<bool>();
+                atmb.SetResult(true);
+                yield return new object[] { new StrongBox<Task>(atmb.Task) };
+            }
         }
     }
 }
