+++ src/libraries/System.Threading/tests/MonitorTests.cs
@@ -2,15 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
 namespace System.Threading.Tests
 {
-    public static partial class MonitorTests
+    public static class MonitorTests
     {
         private const int FailTimeoutMilliseconds = 30000;
 
@@ -437,6 +435,14 @@ namespace System.Threading.Tests
                 Monitor.Pulse(obj);
             }
             Monitor.Exit(obj);
+        }
+
+        [Fact]
+        public static void Enter_HasToWait_LockContentionCountTest()
+        {
+            long initialLockContentionCount = Monitor.LockContentionCount;
+            Enter_HasToWait();
+            Assert.True(Monitor.LockContentionCount - initialLockContentionCount >= 2);
         }
     }
 }
